# Compiled model with torch.compile, unfortunately without performance improvements