--- a/vllm/model_executor/model_loader.py
+++ b/vllm/model_executor/model_loader.py
@@ -85,4 +85,8 @@ def get_model(model_config: ModelConfig, device_config: DeviceConfig,
             # Load the weights from the cached or downloaded files.
             model.load_weights(model_config.model, model_config.download_dir,
                                model_config.load_format, model_config.revision)
-    return model.eval()
+    print('model_config.compile_model', model_config.compile_model)
+    if model_config.compile_model:
+        return torch.compile(model.eval())
+    else:
+        return model.eval()