@UIHCRITT 2023-12-22
using your's code, I run vicuna-7b in one L40, torch.__version__2.1.0+cu121; 
I found 
(flot+compile bs=1,code from gpt-fast) 50.07 tokens/sec;
(float+vllm bs=1) 46.24 tokens/sec; 
(float+vllm+compile) 42.71  tokens/sec;

it seems using torch.compile,without performance improvements;
on the other hand, i try using torch.compile() rewrite model_forward,rather than using @torch.compiler.disable:
in vllm/worker/worker.py  def excute_model

before
''''''
output = self.model(
      input_ids = input_tokens,
      positions = inputs_positions,
      kv_caches=slef.gpu_cache,
      input_metadata=input_metadata,
      cache_events = cache_events,
)
after
'''''''
def _model_foraward(model,input_ids,positions,kv_caches,input_metadata,cache_events):
      return model(input_ids,positions,kv_caches,input_metadata,cache_events)
model_forward = torch.compile(_model_foraward,mode="reduce-overhead",fullgraph=True)
output = model_forward(
      model = self.model,
      input_ids = input_tokens,
      positions = inputs_positions,
      kv_caches=slef.gpu_cache,
      input_metadata=input_metadata,
      cache_events = cache_events,
)
'''''''''''
when I run this code, i get errors''NotImpementedError:ProcessGroupVariable() ''