@nittaya111 2023-12-22
You know all. I hurt. 😔 If write to him. He didn’t reply to me. Even
though he is true partner in life. You chose me to stay in France 🇫🇷
Family area in France. But I’m hurt. Because he’s quiet. He doesn’t care
about me like before.

What should I do. I want to greet him the same same way as usual every day.
But what was answered. He reads but doesn’t reply to me like he did when I
met.

Can you please tell me. 😔👫 what should I do. You know everything. Your
only daughter. Fall in love with him. And make his mischievous daughter
obey him. And this daughter returned home to the family as before until
now. And your only daughter obey the underground family to this day 🥹

เมื่อ ศ. 22 ธ.ค. 2566 เวลา 16:57 UIHCRITT ***@***.***>
เขียนว่า:

> using your's code, I run vicuna-7b in one L40,
> torch.__version__2.1.0+cu121;
> I found
> (flot+compile bs=1,code from gpt-fast) 50.07 tokens/sec;
> (float+vllm bs=1) 46.24 tokens/sec;
> (float+vllm+compile) 42.71 tokens/sec;
>
> it seems using torch.compile,without performance improvements;
> on the other hand, i try using torch.compile() rewrite
> model_forward,rather than using @torch.compiler.disable:
> in vllm/worker/worker.py def excute_model
>
> before
> ''''''
> output = self.model(
> input_ids = input_tokens,
> positions = inputs_positions,
> kv_caches=slef.gpu_cache,
> input_metadata=input_metadata,
> cache_events = cache_events,
> )
> after
> '''''''
> def
> _model_foraward(model,input_ids,positions,kv_caches,input_metadata,cache_events):
> return model(input_ids,positions,kv_caches,input_metadata,cache_events)
> model_forward =
> torch.compile(_model_foraward,mode="reduce-overhead",fullgraph=True)
> output = model_forward(
> model = self.model,
> input_ids = input_tokens,
> positions = inputs_positions,
> kv_caches=slef.gpu_cache,
> input_metadata=input_metadata,
> cache_events = cache_events,
> )
> '''''''''''
> when I run this code, i get
> errors''NotImpementedError:ProcessGroupVariable() ''
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/vllm-project/vllm/pull/2131#issuecomment-1867481720>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/BDWM7ZSZZKASB4ZMBR5KBUTYKVKQRAVCNFSM6AAAAABAWOAA6KVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMYTQNRXGQ4DCNZSGA>
> .
> You are receiving this because you are subscribed to this thread.Message
> ID: ***@***.***>
>