@Lvjinhong 2024-01-12
For the latest version v0.2.7, is there any meaningful acceleration in terms of the compiler?